As Mark did before, we probably don't need a return value from these unref functions. Also, make the ref counts unsigned.
--HG--
extra : convert_revision : e2886fb28589e96d33e8377984790990878f27e7

+++ libpurple/protocols/msn/msg.h
@@ -74,7 +74,7 @@ typedef void (*MsnMsgCb)(MsnMessage *, void *data);
  */
 struct _MsnMessage
 {
-	size_t ref_count;           /**< The reference count.       */
+	guint ref_count;        /**< The reference count.       */
 
 	MsnMsgType type;
 
@@ -89,8 +89,8 @@ struct _MsnMessage
 	char *charset;
 	char *body;
 	gsize body_len;
-	guint total_chunks;   /**< How many chunks in this multi-part message */
-	guint received_chunks; /**< How many chunks we've received so far */
+	guint total_chunks;     /**< How many chunks in this multi-part message */
+	guint received_chunks;  /**< How many chunks we've received so far */
 
 	GHashTable *header_table;
 	GList *header_list;
